services/pedidos/requirements.txt
Flask==1.0.2
Flask-CLI==0.4.0
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.6.1
Flask-Testing==0.7.1
coverage==4.5.3
flake8===3.7.7